+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [unrelased]
+
+### Added
+
+### Fixed
+
+- Fixed make env to append to Lua paths rather than replace them
 
 ### Changed
 
 - Changed machine serialization to use protobuf instead of boost
+- Changed cartesi-machine.lua to save voucher and notice hashes on revert
+- Changed rollup-memory-range.lua to input/output JSON objects
+- Changed to LuaRocks for Lua dependencies
 
 ## [0.8.0] - 2021-12-28